+++ resources/mount.rb
@@ -20,6 +20,9 @@ attribute :path, :kind_of => String, :name_attribute => true
 attribute :cifs_path, :kind_of => String, :required => true
 attribute :username, :kind_of => String, :default => nil
 attribute :password, :kind_of => String, :default => nil
+attribute :dir_owner, :kind_of => String, :default => nil
+attribute :dir_group, :kind_of => String, :default => nil
+attribute :dir_mode, :kind_of => String, :default => nil
 attribute :options, :kind_of => Hash, :default => nil
 
 def initialize( *args )